Замена int на long

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
 
-        SalesManager oSale = new SalesManager(new int[]{});
+        SalesManager oSale = new SalesManager(new long[]{});
         oSale.max();
